Use ProductItemFeature(Unified) in SearchDAO
This solves a huge problem, the largest problem with Products, and it solves it so beautifully: change the FROM part of the query to select from the view. The end. That's it. This would have required a billion joins, this could have been extremely complex, yet it was so simple as we have moved complexity to the views.